requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==45.0.6
dnspython==2.7.0
docutils==0.22
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-mail==1.5.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jwt==1.4.0
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.4
nicegui==2.23.0
orjson==3.11.2
propcache==0.3.2
pscript==0.7.7
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
typer==0.16.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
uvloop==0.21.0
vbuild==0.8.2
watchfiles==1.1.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1

# MQTT and Communication
paho-mqtt==2.1.0

# Backblaze B2 Storage
b2sdk==2.7.0

# gRPC
grpcio==1.70.0
grpcio-tools==1.70.0
protobuf==5.31.1
